From DJRobXs PM5 fork: Merging MJR's nudge filter plus some other VP 9.9.x branch fixes
git-svn-id: https://svn.code.sf.net/p/vpinball/code/branches/9.9.0_PM5_PlungerAccelMOD@2964 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ pin3d.cpp
@@ -806,11 +806,11 @@ void PinProjection::Rotate(const GPINFLOAT x, const GPINFLOAT y, const GPINFLOAT
 {
 	Matrix3D matRotateX, matRotateY, matRotateZ;
 
-	matRotateX.RotateXMatrix(x);
+	matRotateX.RotateXMatrix((float)x);
 	m_matWorld.Multiply(matRotateX, m_matWorld);
-	matRotateY.RotateYMatrix(y);
+	matRotateY.RotateYMatrix((float)y);
 	m_matWorld.Multiply(matRotateY, m_matWorld);
-	matRotateZ.RotateZMatrix(z);
+	matRotateZ.RotateZMatrix((float)z);
 	m_matWorld.Multiply(matRotateZ, m_matWorld);        // matWorld = rotZ * rotY * rotX * origMatWorld
 }
